This version test the framerate performance between SurfaceView and CustomView canvas by measuring FPS.
Also, when the hardware acceleration is disabled for the whole application in the manifest, the canvas.getHeight() returns a diferent value than the same method when the canvas is hardware accelerated. The canvas in the non-hardware accelerated application seems to fill the entire screen and is being pushed down by the title bar. This does not happen though when the hardware acceleration is disabled in the view.